终于都弄好了！脑瓜疼

+++ app/src/main/java/com/giiisp/giiisp/common/ImageviewDouble.java
@@ -2,7 +2,7 @@ package com.giiisp.giiisp.common;
 
 public interface ImageviewDouble {
 
-    void doubleClick(float x,float y,float scale);
+    void doubleClick(float x,float y,int scale);
 
 
 }
